+++ CHANGELOG.md
@@ -4,7 +4,8 @@ CHANGELOG
 ## 2.0.14 (`2001470`)
 
 - **Fix:** Time interval does not apply correctly (#75)
-- **Fix:** Changing post date does not rebuild first/last post info (#84)
+- **Fix:** Changing post date does not rebuild counters on thread (#84)
+- **Fix:** Changing profile post comment does not rebuild counters on profile post (#85)
 
 ## 2.0.13 (`2001370`)